core/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.2
attrs==23.1.0
bardapi==0.1.23a0
blessed==1.20.0
browser-cookie3==0.19.1
cachetools==4.2.4
catalogue==2.0.10
charset-normalizer==3.1.0
chatgpt-cli-tool==1.6.3
contexttimer==0.3.3
decorator==5.1.1
deep-translator==1.11.4
distlib==0.3.7
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
exceptiongroup==1.1.2
executing==2.1.0
filelock==3.12.2
frozenlist==1.3.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.48.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
humanfriendly==10.0
hyperframe==6.0.1
ipython==8.30.0
jedi==0.19.2
joblib==1.4.2
lz4==4.3.2
matplotlib-inline==0.1.7
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
numpy==2.2.0
openai==0.27.7
pandas==2.2.3
parso==0.8.4
periodic==2.1.1
pillow==10.2.0
platformdirs==3.9.1
prompt_toolkit==3.0.48
protobuf==3.20.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
PTable==0.9.2
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycryptodomex==3.18.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-xlib==0.15
pytweening==1.0.7
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==0.7.7
srsly==2.5.0
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.5.0
tk==0.1.0
tqdm==4.65.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
virtualenv==20.24.1
wcwidth==0.2.6
yarl==1.9.2
